pyproject.toml
[tool.poetry]
name = "azeroth-bound-bot"
version = "1.0.0"
description = "Discord bot for WoW Classic+ guild management with webhook-driven architecture"
authors = ["Paweł Kochanowiz <https://github.com/pkochanowicz>"]
readme = "README.md"
packages = [{include = "azeroth_bound_bot", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
gspread = "^6.0.0"
google-auth = "^2.27.0"
google-auth-oauthlib = "^1.2.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = "test_*.py"
